Added a Downloaded state to the ModelFile
The Downloaded state indicates that all remote files for a ModelFile
exist and are larger in size locally. The check is done for all child
files that are found remotely, so any local-only files will not affect
this state. E.g. large local-only files that push the local directory
size above that of the remote directory size will not cause the state
to be Downloaded.