Ng-config: bug fix - config service breaking DI
The config service was in fact NOT breaking dependency injection.
The base web service was creating a subscriber in the constructor,
which called methods on derived class. With DI, this subscriber
was being called right away, even before derived class's ctor was
done. As a result the method being called inside subscriber was
hitting null members.
Fix: subscribe outside the ctor in onInit(). Created a factory and
provider for ConfigService which calls the onInit() method