Py-web: Partial history caching in stream logs
Stream log now returns a cached history of log records from before
the starting of the stream. This allow the stream to connect a bit
late after the start of the app and still be able to receive the
logs and errors that occured during the delay.
History is currently set to 3 seconds.