Py: fixed auto re-extractions on service restarts
After service is restarted, there's a period where model is empty.
The controller removed any file in list of extracted that is not the model.
This clears the extracted list during the period where model is empty.
This causes all matching files to be auto re-extracted.

Fix: don't remove files that are not in model.