Py: fixed 100% CPU utilization even when idle
This issue had two causes:
1. ModelBuilder.build()
When the model is reasonable sized, this can take upwards of 0.5 secs.
Since the controller thread loop sleep interval was also 0.5 secs, this
was always using up the CPU

2. ModelDiffUtil.diff_models
For large models, this can take multiple seconds.

Fix:
The fix for both cases is to cache the model in ModelBuilder and
only do the diff+model updates in controller if the model changed.
The caching in ModelBuilder now takes a look at whether any of the
source data (lftp statuses, local files, remote files) actually
changed, before invalidating the cache.