Ng: cache log records since beginning of the app
Had to remove the interlay of connected/disconnected signals
in the logs. This is because there is no way to preload the component
before the route is clicked. The other alternative is to create another
service which mixes the log records with the connected/disconnected
signals, but that is too much effort for a small convinience.